+++ cmd/bbn/inference.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/mlange-42/bbn"
 	"github.com/mlange-42/bbn/internal/tui"
-	"github.com/mlange-42/bbn/internal/ve"
+	"github.com/mlange-42/bbn/ve"
 	"github.com/spf13/cobra"
 )