+++ internal/tui/node.go
@@ -7,7 +7,7 @@ import (
 	"unicode/utf8"
 
 	"github.com/mlange-42/bbn"
-	"github.com/mlange-42/bbn/internal/ve"
+	"github.com/mlange-42/bbn/ve"
 )
 
 const maxNodeLabelWidth = 48